remove react from integrationsHelpers
