+++ packages/react-mobx/src/context/rattus-context.tsx
@@ -1,6 +1,6 @@
 import type { RattusOrmInstallerOptions } from '@rattus-orm/core/utils/integrationsHelpers'
 import { contextBootstrap } from '@rattus-orm/core/utils/integrationsHelpers'
-import { RattusReactContext } from '@rattus-orm/core/utils/integrationsHelpers'
+import { RattusReactContext } from '@rattus-orm/core/utils/reactIntegrationHelpers'
 import React, { type PropsWithChildren, useMemo } from 'react'
 
 import { ReactMobxDataProvider } from '../data-provider/react-mobx-data-provider'
@@ -14,4 +14,4 @@ export function RattusProvider(props: PropsWithChildren<RattusOrmInstallerOption
   return <RattusReactContext.Provider value={createdDatabase}>{props.children}</RattusReactContext.Provider>
 }
 
-export { reactUseDatabase as useDatabase } from '@rattus-orm/core/utils/integrationsHelpers'
+export { reactUseDatabase as useDatabase } from '@rattus-orm/core/utils/reactIntegrationHelpers'
