+++ packages/react-redux/src/context/rattus-context.tsx
@@ -1,5 +1,6 @@
-import { type RattusOrmInstallerOptions, RattusReactContext } from '@rattus-orm/core/utils/integrationsHelpers'
+import { type RattusOrmInstallerOptions } from '@rattus-orm/core/utils/integrationsHelpers'
 import { contextBootstrap } from '@rattus-orm/core/utils/integrationsHelpers'
+import { RattusReactContext } from '@rattus-orm/core/utils/reactIntegrationHelpers'
 import React, { type PropsWithChildren, useMemo } from 'react'
 import { Provider } from 'react-redux'
 import type { Reducer, Store } from 'redux'
@@ -24,4 +25,4 @@ export function RattusProvider(props: PropsWithChildren<RattusProviderProps>) {
   )
 }
 
-export { reactUseDatabase as useDatabase } from '@rattus-orm/core/utils/integrationsHelpers'
+export { reactUseDatabase as useDatabase } from '@rattus-orm/core/utils/reactIntegrationHelpers'
